+++ Microsoft.Azure.Cosmos/src/ChangeFeedProcessor/Exceptions/FeedSplitException.cs
@@ -11,7 +11,7 @@ namespace Microsoft.Azure.Cosmos.ChangeFeed.Exceptions
     /// Exception occurred during feed processing because of a split.
     /// </summary>
     [Serializable]
-    internal class FeedSplitException : Exception
+    internal class FeedSplitException : FeedException
     {
         /// <summary>
         /// Initializes a new instance of the <see cref="FeedSplitException"/> class using error message and last continuation token.
@@ -19,9 +19,8 @@ namespace Microsoft.Azure.Cosmos.ChangeFeed.Exceptions
         /// <param name="message">The exception error message.</param>
         /// <param name="lastContinuation"> Request continuation token.</param>
         public FeedSplitException(string message, string lastContinuation)
-            : base(message)
+            : base(message, lastContinuation)
         {
-            this.LastContinuation = lastContinuation;
         }
 
         /// <summary>
@@ -32,23 +31,6 @@ namespace Microsoft.Azure.Cosmos.ChangeFeed.Exceptions
         protected FeedSplitException(SerializationInfo info, StreamingContext context)
             : base(info, context)
         {
-            this.LastContinuation = (string)info.GetValue("LastContinuation", typeof(string));
-        }
-
-        /// <summary>
-        /// Gets the value of request continuation token.
-        /// </summary>
-        public string LastContinuation { get; }
-
-        /// <summary>
-        /// Sets the System.Runtime.Serialization.SerializationInfo with information about the exception.
-        /// </summary>
-        /// <param name="info">The SerializationInfo object that holds serialized object data for the exception being thrown.</param>
-        /// <param name="context">The StreamingContext that contains contextual information about the source or destination.</param>
-        public override void GetObjectData(SerializationInfo info, StreamingContext context)
-        {
-            base.GetObjectData(info, context);
-            info.AddValue("LastContinuation", this.LastContinuation);
         }
     }
 }
